Extracted variables to separate files.
