+++ src/PerformanceTests/Tests/Categories/AuditOnVsOffFixture.cs
@@ -19,7 +19,7 @@ namespace Categories
             return PermutationGenerator.Generate(new Permutations
             {
                 Versions = new[] { NServiceBusVersion.V5, NServiceBusVersion.V6, },
-                IOPS = new[] { IOPS.Normal },
+                IOPS = new[] { IOPS.Default },
                 Platforms = new[] { Platform.x86, },
                 GarbageCollectors = new[] { GarbageCollector.Client, },
                 Transports = new[] { Transport.MSMQ },
