+++ src/PerformanceTests/Tests/Categories/PublishVsSendFixture.cs
@@ -33,7 +33,7 @@ namespace Categories
             return PermutationGenerator.Generate(new Permutations
             {
                 Versions = new[] { NServiceBusVersion.V5, NServiceBusVersion.V6, },
-                IOPS = new[] { IOPS.Normal },
+                IOPS = new[] { IOPS.Default },
                 Platforms = new[] { Platform.x86, },
                 GarbageCollectors = new[] { GarbageCollector.Client, },
                 Transports = new[] { Transport.MSMQ, Transport.ASB, Transport.ASQ, Transport.MSMQ, Transport.RMQ, Transport.SQL,  },
