+++ src/PerformanceTests/Tests/Variables/Audit.cs
@@ -1,0 +1,8 @@
+namespace Variables
+{
+    public enum Audit
+    {
+        On,
+        Off
+    }
+}
