+++ src/PerformanceTests/Tests/Variables/ConcurrencyLevel.cs
@@ -1,0 +1,9 @@
+namespace Variables
+{
+    public enum ConcurrencyLevel
+    {
+        Sequential,
+        EnvCores,
+        EnvCores4x
+    }
+}
