+++ src/PerformanceTests/Tests/Variables/DTC.cs
@@ -1,0 +1,8 @@
+namespace Variables
+{
+    public enum DTC
+    {
+        On,
+        Off
+    }
+}
