+++ src/PerformanceTests/Tests/Variables/GarbageCollector.cs
@@ -1,0 +1,8 @@
+namespace Variables
+{
+    public enum GarbageCollector
+    {
+        Client,
+        Server
+    }
+}
