+++ src/PerformanceTests/Tests/Variables/IOPS.cs
@@ -1,0 +1,10 @@
+namespace Variables
+{
+    public enum IOPS
+    {
+        Default
+        //Slow, // HDD
+        //Normal, // SSD
+        //Fast // SSD M2
+    }
+}
