+++ src/PerformanceTests/Tests/Variables/MessageSize.cs
@@ -1,0 +1,10 @@
+namespace Variables
+{
+    public enum MessageSize
+    {
+        Tiny,
+        Small,
+        Medium,
+        Large
+    }
+}
