+++ src/PerformanceTests/Tests/Variables/Outbox.cs
@@ -1,0 +1,8 @@
+namespace Variables
+{
+    public enum Outbox
+    {
+        On,
+        Off
+    }
+}
