+++ src/PerformanceTests/Tests/Variables/Persistence.cs
@@ -1,0 +1,10 @@
+namespace Variables
+{
+    public enum Persistence
+    {
+        InMemory,
+        NHibernate,
+        RavenDB,
+        //MSMQ
+    }
+}
