+++ src/PerformanceTests/Tests/Variables/Platform.cs
@@ -1,0 +1,8 @@
+﻿namespace Variables
+{
+    public enum Platform
+    {
+        x86,
+        x64
+    }
+}
