+++ src/PerformanceTests/Tests/Variables/Serialization.cs
@@ -1,0 +1,8 @@
+namespace Variables
+{
+    public enum Serialization
+    {
+        Json,
+        Xml
+    }
+}
