+++ src/PerformanceTests/Tests/Variables/TransactionMode.cs
@@ -1,0 +1,10 @@
+namespace Variables
+{
+    public enum TransactionMode
+    {
+        Default,
+        Receive,
+        Unreliable,
+        AtomicSends
+    }
+}
