+++ src/PerformanceTests/Tests/Variables/Transport.cs
@@ -1,0 +1,11 @@
+namespace Variables
+{
+    public enum Transport
+    {
+        MSMQ,
+        RMQ,
+        ASQ,
+        ASB,
+        SQL
+    }
+}
